added new functions in bechmark example
